requirements.txt.txt
tensorflow-gpu==1.15
numpy
os
scipy
matplotlib
time
random
shutil
datetime